chore: introduce TRG category based package structure
